pyproject.toml
[tool.poetry]
name = "shanghaicopyright"
version = "0.1.0"
description = ""
authors = ["Silverowo <1581647952@qq.com>, nullqwertyuiop <nullqwertyuiop@qq.com>, Little-LinNian <2544704967@qq.com>"]
license = "AGPL-3.0 license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.3"
pydantic = "^1.10.4"
contextvars = "^2.4"
asyncio = "^3.4.3"
PyYAML = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
